Add Get feedback endpoints

+++ Service/FeedbackService.cs
@@ -3,7 +3,9 @@ using Contracts;
 using Entities.Exceptions;
 using Entities.Models;
 using Service.Contracts;
+using Shared.DTO;
 using Shared.DTO.Feedback;
+using Shared.RequestFeatures;
 
 namespace Service
 {
@@ -49,12 +51,8 @@ namespace Service
 
 		}
 
-		public async  Task DeleteFeedbackAsync(string CompanyID, int PostId, string StudentId, FeedbackDelationDto FeedbackId)
+		public async  Task DeleteFeedbackAsync(string CompanyID, int PostId, FeedbackDelationDto FeedbackId)
 		{
-			var student = _repository.Student.GetStudent(StudentId, false);
-			if (student == null)
-				throw new StudentNotFoundException(StudentId);
-
 			object post;
 			if (FeedbackId.Type == FeedbackType.Scholarship)
 			{
@@ -74,9 +72,50 @@ namespace Service
 			if (feedback == null)
 				throw new FeedbackNotFoundException(FeedbackId.Id);
 
+			if (feedback.PostId != PostId)
+				throw new BadRequestException("Invalid data");
+
 			_repository.FeedbackRepository.DeleteFeedback(feedback);
 
 			await _repository.SaveAsync();
+		}
+
+		public async Task<FeedBackDto> GetFeedbackAsync(int FeedbackId)
+		{
+
+			var feedback = await _repository.FeedbackRepository.GetFeedbackById(FeedbackId, false);
+
+
+			if (feedback == null)
+				throw new FeedbackNotFoundException(FeedbackId);
+
+			var res =_mapper.Map<FeedBackDto>(feedback);
+			return res;
+		}
+
+
+		public async Task<PagedList<FeedBackDto>> GetAllFeedbackAsync(string CompanyID, int PostId, FeedBackParameters Feedback)
+		{
+
+			object post;
+			if (Feedback.type == FeedbackType.Scholarship)
+			{
+				post = _repository.Scholarship.GetScholarship(CompanyID, PostId, false);
+				if (post == null)
+					throw new ScholarshipNotFoundException(PostId);
+			}
+			else
+			{
+				post = _repository.Intership.GetInternshipById(PostId, false);
+				if (post == null)
+					throw new InternshipNotFoundException(PostId);
+			}
+
+			var feedbacks = await _repository.FeedbackRepository.GetAllFeedbackAsync(PostId,Feedback, false);
+
+			var res = _mapper.Map<List<FeedBackDto>>(feedbacks);
+
+			return new PagedList<FeedBackDto>(res, feedbacks.MetaData.TotalCount, Feedback.PageNumber, Feedback.PageSize);
 		}
 	}
 }
